+++ packages/headless/src/features/debug/ranking-info-parser.ts
@@ -1,44 +1,160 @@
-export type RankingInfo = ReturnType<typeof parseRankingInfo>;
+export interface RankingInformation {
+  /**
+   * The attributes of the document that contributed to its ranking.
+   * */
+  documentWeights: DocumentWeights | null;
 
-export interface ListOfWeights {
+  /**
+   * The weight attributed to each term in the query.
+   */
+  termsWeight: TermWeightReport | null;
+
+  /**
+   * The sum of all weights.
+   */
+  totalWeight: number | null;
+
+  /**
+   * The weights applied by query ranking expressions.
+   */
+  qreWeights: QueryRankingExpressionWeights[];
+}
+
+export interface DocumentWeights {
+  /**
+   * The effect of proximity of query terms in the document. More weight is given to documents having the terms closer together.
+   */
   Adjacency: number;
+
+  /**
+   * The weight attributed to user ratings.
+   */
   'Collaborative rating': number;
+
+  /**
+   * The weight assigned through an [indexing pipeline extension (IPE)](https://docs.coveo.com/en/206/) for the item.
+   */
   Custom: number;
+
+  /**
+   * The weight attributed to the document date.
+   */
   Date: number;
+
+  /**
+   * The weight applied by a [query ranking expression (QRE)](https://docs.coveo.com/en/2777/).
+   */
   QRE: number;
+
+  /**
+   * The weight attributed to rank of the document in the documents remaining after filtering indexed items by query terms and user permissions.
+   * See [item weighting](https://docs.coveo.com/en/1624/#phase-2-item-weighting) for more information.
+   */
   Quality: number;
+
+  /**
+   * The weight applied by a [ranking function](https://docs.coveo.com/en/1448/).
+   */
   'Ranking functions': number;
+
+  /**
+   * The effect of the reputation of the document source on the ranking.
+   */
   Source: number;
+
+  /**
+   * The weight attributed to presence of query terms in the document title.
+   */
   Title: number;
+
+  /**
+   * Custom factors affecting the document weight.
+   */
   [key: string]: number;
 }
 
-export interface ListOfQRE {
+export interface QueryRankingExpressionWeights {
+  /**
+   * The query ranking expression (QRE).
+   */
   expression: string;
+
+  /**
+   * The score added by the query ranking expression (QRE).
+   */
   score: number;
 }
 
-export type ListOfTermsWeights = Record<string, WeightsPerTerm>;
+export type TermWeightReport = Record<string, StemmedTermInformation>;
 
-export interface WeightsPerTerm {
-  Weights: WeightsPerTermBreakdown;
-  terms: Record<string, WeightsPerTermPerDocument>;
+export interface StemmedTermInformation {
+  Weights: TermWeights | null;
+  terms: Record<string, TermWeightsPerDocument>;
 }
 
-export interface WeightsPerTermBreakdown {
+interface TermWeights {
+  /**
+   * The weight allocated when query terms have a special casing in the document.
+   */
   Casing: number;
+
+  /**
+   * The weight allocated to the presence of query terms in the automatically populated '@concepts' field of the document.
+   */
   Concept: number;
+
+  /**
+   * The weight allocated when query terms are formatted in the document (e.g., heading level, bold, large, etc.).
+   */
   Formatted: number;
+
+  /**
+   * The weight allocated based on the number of times query terms appear in the document.
+   */
   Frequency: number;
+
+  /**
+   * The weight allocated when the document contains words with the same root as the query terms.
+   *
+   * @example
+   * Searching for `programmer` will match documents with `programmer`, `programmers`, `program`, `programming`, etc.
+   */
   Relation: number;
+
+  /**
+   * The weight allocated when the document summary contains query terms.
+   */
   Summary: number;
+
+  /**
+   * The weight allocated when the document title contains query terms.
+   */
   Title: number;
+
+  /**
+   * The weight allocated when the document URI contains query terms.
+   */
   URI: number;
+
+  /**
+   * Custom factors affecting the term weight.
+   */
   [key: string]: number;
 }
 
-export interface WeightsPerTermPerDocument {
+interface TermWeightsPerDocument {
+  /**
+   * Captures the likelihood that query term expansions are related to the original query term. Documents containing highly correlated expansions are ranked higher than ones containing poorly correlated expansions.
+   *
+   * @example
+   *
+   * When you search for `universe`, because of the way the stemming algorithm works, the index expands the query using terms from the `univer` stem classes that can include `university`. When the terms `universe` and `university` rarely co-occur in your indexed items, items containing `university` are ranked lower.
+   */
   Correlation: number;
+
+  /**
+   * The number of times a queried keyword appears in a given item, offset by the number of items in the index containing that keyword (see [TF-IDF](https://en.wikipedia.org/wiki/Tf%E2%80%93idf)).
+   */
   'TF-IDF': number;
 }
 
@@ -56,7 +172,7 @@ export const parseRankingInfo = (value: string) => {
   const totalWeigthRegexResult = REGEX_EXTRACT_TOTAL_WEIGHTS.exec(value);
 
   const qreWeights = parseQREWeights(value);
-  const documentWeights = parseWeights(
+  const documentWeights = parseWeights<DocumentWeights>(
     docWeightsRegexResult ? docWeightsRegexResult[1] : null
   );
   const termsWeight = parseTermsWeights(termsWeightRegexResult);
@@ -72,7 +188,9 @@ export const parseRankingInfo = (value: string) => {
   };
 };
 
-const parseWeights = (value: string | null): ListOfWeights | null => {
+const parseWeights = <T extends Record<string, number>>(
+  value: string | null
+): T | null => {
   const REGEX_EXTRACT_LIST_OF_WEIGHTS = /(\w+(?:\s\w+)*): ([-0-9]+)/g;
   const REGEX_EXTRACT_WEIGHT_GROUP = /^(\w+(?:\s\w+)*): ([-0-9]+)$/;
 
@@ -86,7 +204,7 @@ const parseWeights = (value: string | null): ListOfWeights | null => {
     return null;
   }
 
-  const weights = {} as ListOfWeights;
+  const weights: Record<string, number> = {};
 
   for (const weight of listOfWeight) {
     const weightGroup = weight.match(REGEX_EXTRACT_WEIGHT_GROUP);
@@ -97,7 +215,7 @@ const parseWeights = (value: string | null): ListOfWeights | null => {
       weights[weightAppliedOn] = Number(weightValue);
     }
   }
-  return weights;
+  return weights as T;
 };
 
 const matchExec = (value: string, regex: RegExp) => {
@@ -111,7 +229,7 @@ const matchExec = (value: string, regex: RegExp) => {
 
 const parseTermsWeights = (
   termsWeight: RegExpExecArray | null
-): ListOfTermsWeights | null => {
+): TermWeightReport | null => {
   const REGEX_EXTRACT_GROUP_OF_TERMS = /((?:[^:]+: [0-9]+, [0-9]+; )+)\n((?:\w+: [0-9]+; )+)/g;
   const REGEX_EXTRACT_SINGLE_TERM = /([^:]+): ([0-9]+), ([0-9]+); /g;
 
@@ -123,11 +241,11 @@ const parseTermsWeights = (
   if (!listOfTerms) {
     return null;
   }
-  const terms = {} as ListOfTermsWeights;
+  const terms: TermWeightReport = {};
   for (const term of listOfTerms) {
     const listOfWords = matchExec(term[1], REGEX_EXTRACT_SINGLE_TERM);
 
-    const words = {} as Record<string, WeightsPerTermPerDocument>;
+    const words: Record<string, TermWeightsPerDocument> = {};
     for (const word of listOfWords) {
       words[word[1]] = {
         Correlation: Number(word[2]),
@@ -135,22 +253,22 @@ const parseTermsWeights = (
       };
     }
 
-    const weights = parseWeights(term[2]);
+    const weights = parseWeights<TermWeights>(term[2]);
     terms[Object.keys(words).join(', ')] = {
       terms: words,
-      Weights: (weights as unknown) as WeightsPerTermBreakdown,
+      Weights: weights,
     };
   }
 
   return terms;
 };
 
-const parseQREWeights = (value: string): ListOfQRE[] => {
+const parseQREWeights = (value: string): QueryRankingExpressionWeights[] => {
   const REGEX_EXTRACT_QRE_WEIGHTS = /(Expression:\s".*")\sScore:\s(?!0)([0-9]+)\n+/g;
 
   let qreWeightsRegexResult = REGEX_EXTRACT_QRE_WEIGHTS.exec(value);
 
-  const qreWeights: ListOfQRE[] = [];
+  const qreWeights: QueryRankingExpressionWeights[] = [];
   while (qreWeightsRegexResult) {
     qreWeights.push({
       expression: qreWeightsRegexResult[1],